feat: Needleman-Wunsch global alignment algorithm (#2)
Implements the Needleman-Wunsch algorithm for finding the longest path
in a DAG.

As this is a bioinformatics course, we know that the DAG will be a grid
shape and so I exploit that when looking at the possible predecessors,
as well as in the `alignmentGrid` struct.
The `alignmentGrid` struct uses a 1D row-major slice to store all
possible values in the grid; there are helper methods for getting and
setting here to ease the indexing calculation.
There is a simple `String() string` method on `alignmentGrid` for
visualising small examples.