feat: UPGMA algorithm for phylogenetic tree construction (#5)
Add an implementation of the UPGMA algorithm for phylogenetic tree
construction.
Rather than 'deleting' items from the matrix once they have been merged
into other clusters, this implementation maintains a list of
`deadClusters` that are skipped by the algorithm.
Lookup into deadClusters is currently O(n), but since n is small this is
OK. It could be made faster with a `map[int]struct{}`.

This PR also includes pretty printing methods on UPGMA trees, although
the majority of that code is courtesy of Codex, to save time.

A follow-up PR could also add a better entry point for the algorithm,
rather than just running a hard-coded example.